refactor: move some parts in lib.rs to realtime.rs
